api/flash: Move the rebooting of keyboards to api/focus
api/focus is in a far better position to handle reboots than api/flash - or
FlashCommands - is, so move the rebooting there. This makes the code simpler,
and fixes a bug where we weren't properly rebooting the keyboard.

We weren't rebooting the keyboard because `focus.request("device.reset")` did
not actually run. It did not run, because `supported_commands()` was returning
undefined (it happens sometimes, under circumstances I have not dived into yet).
Basically, because `focus.request()` does some sanity checking which we do *not*
need in this case.

By moving the rebooting to focus, we can use the internal `_request()` method,
which doesn't do any of that.

Fixes #1154.
